refactor - moved actions to correct files and moved common logic to playbooks_utils
